Addressing iOS7 crashing bug
I think this should address the browser crash issue on iOS7.  I created
a new page called index_datalist.php and changed the chosen selector to
a <datalist> selector on that page.  Then I added a script to the
original index.php to redirect to index_datalist.php.

index_datalist.php points to the new main_datalist.js which has
rewritten js that grabs the info from the datalist and passes it to the
php function that draws the trips.

Hope this works.